Move Governance and sharing widget creation into EntityMetadata
Clean compilation.
