+++ src/main/java/org/sagebionetworks/web/client/widget/entity/EntityMetadataView.java
@@ -2,6 +2,8 @@ package org.sagebionetworks.web.client.widget.entity;
 
 import org.sagebionetworks.repo.model.VersionInfo;
 import org.sagebionetworks.web.client.model.EntityBundle;
+import org.sagebionetworks.web.client.utils.APPROVAL_REQUIRED;
+import org.sagebionetworks.web.client.utils.Callback;
 import org.sagebionetworks.web.shared.PaginatedResults;
 
 import com.google.gwt.user.client.rpc.AsyncCallback;
@@ -15,10 +17,46 @@ public interface EntityMetadataView extends IsWidget {
 
 	public void setReadOnly(boolean readOnly);
 
+	public void showInfo(String string, String message);
+
 	public interface Presenter {
 
 		void loadVersions(String id, int offset, int limit,
 				AsyncCallback<PaginatedResults<VersionInfo>> asyncCallback);
+		boolean isAnonymous();
+
+		/**
+		 *
+		 * @return
+		 * @exception if anonymous
+		 */
+		String getJiraFlagUrl();
+
+		/**
+		 *
+		 */
+		String getJiraRequestAccessUrl();
+
+		boolean hasAdministrativeAccess();
+
+		boolean includeRestrictionWidget();
+
+		public APPROVAL_REQUIRED getRestrictionLevel();
+
+		Callback accessRequirementCallback();
+
+		Callback getImposeRestrictionsCallback();
+
+		boolean hasFulfilledAccessRequirements();
+
+		String accessRequirementText();
+
+		boolean isTermsOfUseAccessRequirement();
+
+		Callback getLoginCallback();
+
+		void fireEntityUpdatedEvent();
 
 	}
+
 }
