+++ src/main/java/org/sagebionetworks/web/client/widget/entity/EntityPageTopViewImpl.java
@@ -1,7 +1,6 @@
 package org.sagebionetworks.web.client.widget.entity;
 
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 
 import org.sagebionetworks.repo.model.Entity;
@@ -10,8 +9,6 @@ import org.sagebionetworks.repo.model.Folder;
 import org.sagebionetworks.repo.model.Project;
 import org.sagebionetworks.repo.model.Summary;
 import org.sagebionetworks.repo.model.UserProfile;
-import org.sagebionetworks.repo.model.VersionInfo;
-import org.sagebionetworks.repo.model.Versionable;
 import org.sagebionetworks.repo.model.attachment.AttachmentData;
 import org.sagebionetworks.repo.model.attachment.UploadResult;
 import org.sagebionetworks.repo.model.attachment.UploadStatus;
@@ -24,9 +21,6 @@ import org.sagebionetworks.web.client.SynapseJSNIUtils;
 import org.sagebionetworks.web.client.events.EntityUpdatedEvent;
 import org.sagebionetworks.web.client.events.EntityUpdatedHandler;
 import org.sagebionetworks.web.client.model.EntityBundle;
-import org.sagebionetworks.web.client.utils.APPROVAL_REQUIRED;
-import org.sagebionetworks.web.client.utils.Callback;
-import org.sagebionetworks.web.client.utils.TOOLTIP_POSITION;
 import org.sagebionetworks.web.client.widget.breadcrumb.Breadcrumb;
 import org.sagebionetworks.web.client.widget.entity.children.EntityChildBrowser;
 import org.sagebionetworks.web.client.widget.entity.dialog.AddAttachmentDialog;
@@ -37,7 +31,6 @@ import org.sagebionetworks.web.shared.PaginatedResults;
 import com.extjs.gxt.ui.client.Style.Direction;
 import com.extjs.gxt.ui.client.Style.HorizontalAlignment;
 import com.extjs.gxt.ui.client.data.BaseModelData;
-import com.extjs.gxt.ui.client.data.BasePagingLoadConfig;
 import com.extjs.gxt.ui.client.data.BasePagingLoadResult;
 import com.extjs.gxt.ui.client.data.BasePagingLoader;
 import com.extjs.gxt.ui.client.data.ModelData;
@@ -45,9 +38,6 @@ import com.extjs.gxt.ui.client.data.PagingLoadConfig;
 import com.extjs.gxt.ui.client.data.PagingLoadResult;
 import com.extjs.gxt.ui.client.data.RpcProxy;
 import com.extjs.gxt.ui.client.event.ButtonEvent;
-import com.extjs.gxt.ui.client.event.Events;
-import com.extjs.gxt.ui.client.event.GridEvent;
-import com.extjs.gxt.ui.client.event.Listener;
 import com.extjs.gxt.ui.client.event.SelectionListener;
 import com.extjs.gxt.ui.client.fx.FxConfig;
 import com.extjs.gxt.ui.client.store.ListStore;
@@ -75,21 +65,17 @@ import com.google.gwt.cell.client.widget.PreviewDisclosurePanel;
 import com.google.gwt.core.client.GWT;
 import com.google.gwt.event.dom.client.ClickEvent;
 import com.google.gwt.event.dom.client.ClickHandler;
-import com.google.gwt.i18n.client.DateTimeFormat;
 import com.google.gwt.safehtml.shared.SafeHtml;
 import com.google.gwt.safehtml.shared.SafeHtmlBuilder;
 import com.google.gwt.safehtml.shared.SafeHtmlUtils;
 import com.google.gwt.uibinder.client.UiBinder;
 import com.google.gwt.uibinder.client.UiField;
-import com.google.gwt.user.client.Window;
 import com.google.gwt.user.client.rpc.AsyncCallback;
 import com.google.gwt.user.client.ui.AbstractImagePrototype;
 import com.google.gwt.user.client.ui.Anchor;
 import com.google.gwt.user.client.ui.Composite;
 import com.google.gwt.user.client.ui.HTML;
 import com.google.gwt.user.client.ui.HTMLPanel;
-import com.google.gwt.user.client.ui.Hyperlink;
-import com.google.gwt.user.client.ui.InlineLabel;
 import com.google.gwt.user.client.ui.SimplePanel;
 import com.google.gwt.user.client.ui.Widget;
 import com.google.inject.Inject;
@@ -320,75 +306,6 @@ public class EntityPageTopViewImpl extends Composite implements EntityPageTopVie
 		// ** FULL WIDTH **
 		// none.
 	}
-
-	private Widget createShareSettingsWidget(boolean isPublic) {
-		final SimplePanel lc = new SimplePanel();
-		String styleName = isPublic ? "public-acl-image" : "private-acl-image";
-		String description = isPublic ? DisplayConstants.PUBLIC_ACL_ENTITY_PAGE : DisplayConstants.PRIVATE_ACL_ENTITY_PAGE;
-		String tooltip = isPublic ? DisplayConstants.PUBLIC_ACL_DESCRIPTION : DisplayConstants.PRIVATE_ACL_DESCRIPTION;
-		
-		SafeHtmlBuilder shb = new SafeHtmlBuilder();
-		shb.appendHtmlConstant("<span style=\"margin-right: 5px;\">Sharing:</span><div class=\"" + styleName+ "\" style=\"display:inline-block; position:absolute\"></div>");
-		shb.appendHtmlConstant("<span style=\"margin-right: 10px; margin-left: 20px;\">"+description+"</span>");
-		
-		//form the html
-		HTMLPanel htmlPanel = new HTMLPanel(shb.toSafeHtml());
-		htmlPanel.addStyleName("inline-block");
-		DisplayUtils.addTooltip(synapseJSNIUtils, htmlPanel, tooltip, TOOLTIP_POSITION.RIGHT);
-		lc.add(htmlPanel);
-		
-		return lc;
-	}
-	
-	private Widget createRestrictionWidget() {
-		if (!presenter.includeRestrictionWidget()) return null;
-		boolean isAnonymous = presenter.isAnonymous();
-		boolean hasAdministrativeAccess = false;
-		boolean hasFulfilledAccessRequirements = false;
-		String jiraFlagLink = null;
-		if (!isAnonymous) {
-			hasAdministrativeAccess = presenter.hasAdministrativeAccess();
-			jiraFlagLink = presenter.getJiraFlagUrl();
-		}
-		APPROVAL_REQUIRED restrictionLevel = presenter.getRestrictionLevel();
-		String accessRequirementText = null;
-		Callback touAcceptanceCallback = null;
-		Callback requestACTCallback = null;
-		Callback imposeRestrictionsCallback = presenter.getImposeRestrictionsCallback();
-		Callback loginCallback = presenter.getLoginCallback();
-		if (restrictionLevel!=APPROVAL_REQUIRED.NONE) {
-			accessRequirementText = presenter.accessRequirementText();
-			if (restrictionLevel==APPROVAL_REQUIRED.LICENSE_ACCEPTANCE) {
-				touAcceptanceCallback = presenter.accessRequirementCallback();
-			} else {
-				// get the Jira link for ACT approval
-				if (!isAnonymous) {
-					requestACTCallback = new Callback() {
-						@Override
-						public void invoke() {
-							Window.open(presenter.getJiraRequestAccessUrl(), "_blank", "");
-
-						}
-					};
-				}
-			}
-			if (!isAnonymous) hasFulfilledAccessRequirements = presenter.hasFulfilledAccessRequirements();
-		}
-		return EntityViewUtils.createRestrictionsWidget(
-				jiraFlagLink, 
-				isAnonymous, 
-				hasAdministrativeAccess,
-				accessRequirementText,
-				touAcceptanceCallback,
-				requestACTCallback,
-				imposeRestrictionsCallback,
-				loginCallback,
-				restrictionLevel, 
-				hasFulfilledAccessRequirements,
-				iconsImageBundle,
-				synapseJSNIUtils);
-	}
-
 
 	private void renderSnapshotEntity(EntityBundle bundle, UserProfile userProfile,
 			String entityTypeDisplay, boolean canEdit, boolean readOnly, MarginData widgetMargin) {
